debate_platform/requirements.txt
django==5.2.3
python-dotenv==1.0.0
sarvamai==0.1.0
Pillow==10.0.0
